pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Tata Pestova <tata.pestova@inbox.ru>"]
readme = "README.md"
packages = [{include = "hexlet_code"}]

[tool.poetry.dependencies]
python = "^3.8.1"
django = "^4.2.2"
load-dotenv = "^0.1.0"
gunicorn = "^20.1.0"
django-bootstrap4 = "^23.1"
django-filter = "^23.2"
rollbar = "^0.16.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
coverage = "^7.2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
